+++ src/main/java/roomescape/AdminController.java
@@ -1,0 +1,18 @@
+package roomescape;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class AdminController {
+
+    @GetMapping("/admin")
+    public String homePage() {
+        return "admin/index";
+    }
+
+    @GetMapping("/admin/reservation")
+    public String manageReservation() {
+        return "admin/reservation-legacy";
+    }
+}
